Fix PR #155 review blockers: import structure, monkeypatch target, personal URLs
- Move BLOCKED_BANKS and COOKIES_DIR to public_portfolio/constants.py
- Extract run_session_saver and helpers into public_portfolio/session.py
- runner.py now imports from lib modules only (no bare top-level import)
- browser.py imports COOKIES_DIR from constants instead of redefining it
- save_bank_session.py reduced to thin CLI wrapper
- Fix monkeypatch target in session retry test: load_config -> load_effective_config
- Clear personal contributor URLs from public_portfolios.defaults.json
