+++ markphotomediaapprovalstatus/markphotomediaapprovalstatuslib/public_portfolio/constants.py
@@ -5,6 +5,9 @@ from pathlib import Path
 DEFAULT_PUBLIC_PORTFOLIO_CONFIG = "public_portfolios.json"
 DEFAULT_PUBLIC_PORTFOLIO_DEFAULTS = str(Path(__file__).resolve().parents[2] / "public_portfolios.defaults.json")
 
+BLOCKED_BANKS = ["ShutterStock", "Pond5", "BigStockPhoto", "Dreamstime"]
+COOKIES_DIR = Path(__file__).resolve().parents[2] / "cookies"
+
 PUBLIC_PORTFOLIO_PAGE_TIMEOUT_MS = 120000
 PUBLIC_PORTFOLIO_INITIAL_WAIT_MS = 10000
 PUBLIC_PORTFOLIO_SCROLL_WAIT_MS = 5000
