+++ markphotomediaapprovalstatus/public_portfolios.defaults.json
@@ -1,31 +1,31 @@
 {
   "banks": {
     "ShutterStock": {
-      "portfolio_url": "https://www.shutterstock.com/cs/g/Danny+JP?sort=newest"
+      "portfolio_url": null
     },
     "AdobeStock": {
-      "portfolio_url": "https://stock.adobe.com/cz/contributor/210540880/dan?load_type=author&prev_url=detail"
+      "portfolio_url": null
     },
     "Dreamstime": {
-      "portfolio_url": "https://www.dreamstime.com/dannjp_info"
+      "portfolio_url": null
     },
     "DepositPhotos": {
-      "portfolio_url": "https://depositphotos.com/portfolio-55998584.html"
+      "portfolio_url": null
     },
     "123RF": {
-      "portfolio_url": "https://cz.123rf.com/profile_dannyjpn"
+      "portfolio_url": null
     },
     "Pond5": {
-      "portfolio_url": "https://www.pond5.com/artist/dannyjpn?tab=photo&pp=1&sb=6"
+      "portfolio_url": null
     },
     "GettyImages": {
-      "portfolio_url": "https://www.istockphoto.com/en/portfolio/danielkrupa?assettype=image&mediatype=photography&sort=newest"
+      "portfolio_url": null
     },
     "Alamy": {
       "portfolio_url": null
     },
     "BigStockPhoto": {
-      "portfolio_url": "https://www.bigstockphoto.com/cs/search/?contributor=DannyJPN&order=new"
+      "portfolio_url": null
     },
     "CanStockPhoto": {
       "portfolio_url": null
